backend/requirements.txt
# Django
django==2.2.13
djangorestframework==3.12.4

# Network & Utils
requests==2.23.0

# Cryptography & Security
pyjwt==1.7.1
jwcrypto==0.7

# Misc
six==1.15.0
django-cors-headers==3.7.0
